+++ packages/commerce-sdk-react/src/auth/index.test.ts
@@ -34,6 +34,11 @@ jest.mock('commerce-sdk-isomorphic', () => {
     }
 })
 
+jest.mock('../utils', () => ({
+    __esModule: true,
+    onClient: () => true
+}))
+
 const config = {
     clientId: 'clientId',
     organizationId: 'organizationId',
@@ -89,7 +94,8 @@ describe('Auth', () => {
             id_token: 'id_token',
             idp_access_token: 'idp_access_token',
             token_type: 'token_type',
-            usid: 'usid'
+            usid: 'usid',
+            customer_type: 'guest'
         }
         const {refresh_token_guest, ...result} = {...sample, refresh_token: 'refresh_token_guest'}
 
@@ -110,6 +116,17 @@ describe('Auth', () => {
         expect(auth.isTokenExpired(JWTExpired)).toBe(true)
         // @ts-expect-error private method
         expect(() => auth.isTokenExpired()).toThrow()
+    })
+    test('site switch clears auth storage', () => {
+        const auth = new Auth(config)
+        // @ts-expect-error private method
+        auth.set('access_token', '123')
+        // @ts-expect-error private method
+        auth.set('refresh_token_guest', '456')
+        const switchSiteConfig = {...config, siteId: 'another site'}
+        const newAuth = new Auth(switchSiteConfig)
+        expect(newAuth.get('access_token')).not.toBe('123')
+        expect(newAuth.get('refresh_token_guest')).not.toBe('456')
     })
     test('isTokenExpired', () => {
         const auth = new Auth(config)
@@ -133,7 +150,8 @@ describe('Auth', () => {
             id_token: 'id_token',
             idp_access_token: 'idp_access_token',
             token_type: 'token_type',
-            usid: 'usid'
+            usid: 'usid',
+            customer_type: 'guest'
         }
         // @ts-expect-error private method
         auth.pendingToken = Promise.resolve(data)
@@ -152,7 +170,8 @@ describe('Auth', () => {
             id_token: 'id_token',
             idp_access_token: 'idp_access_token',
             token_type: 'token_type',
-            usid: 'usid'
+            usid: 'usid',
+            customer_type: 'guest'
         }
         const {refresh_token_guest, ...result} = {...data, refresh_token: 'refresh_token_guest'}
 
@@ -175,7 +194,8 @@ describe('Auth', () => {
             id_token: 'id_token',
             idp_access_token: 'idp_access_token',
             token_type: 'token_type',
-            usid: 'usid'
+            usid: 'usid',
+            customer_type: 'guest'
         }
         const {refresh_token_guest, ...result} = {...data, refresh_token: 'refresh_token_guest'}