+++ packages/commerce-sdk-react/src/hooks/ShopperLogin/helper.ts
@@ -4,11 +4,8 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {ShopperLoginTypes} from 'commerce-sdk-isomorphic'
-import {Argument} from '../types'
 import {useMutation} from '../useMutation'
 import useAuth from '../useAuth'
-import Auth from '../../auth'
 import {UseMutationResult} from '@tanstack/react-query'
 
 export const ShopperLoginHelpers = {